TheFuzz v0.2.1 → Overview → TheFuzz → Phonetic → MetaphoneMetric

TheFuzz.Phonetic.MetaphoneMetric

Summary

Functions

Calculates the Metaphone Phonetic Algorithm metric of two strings.

Source

Summary

compare(a, b)
Compares two values phonetically and returns a boolean of whether they match or not. ## Examples

iex> TheFuzz.Phonetic.MetaphoneMetric.compare("Colorado", "Kolorado")
true
iex> TheFuzz.Phonetic.MetaphoneMetric.compare("Moose", "Elk")
false

Functions

compare(a, b)
(function) # ↑

Compares two values phonetically and returns a boolean of whether they match or not. ## Examples

iex> TheFuzz.Phonetic.MetaphoneMetric.compare("Colorado", "Kolorado")
true
iex> TheFuzz.Phonetic.MetaphoneMetric.compare("Moose", "Elk")
false

Source