TheFuzz v0.1.0

Modules

Modules summary

TheFuzz
Contains shortforms to execute different string metric algorithms to compare given strings

TheFuzz.Phonetic.MetaphoneAlgorithm
Calculates the Metaphone Phonetic Algorithm of a string

TheFuzz.Phonetic.MetaphoneMetric
Calculates the Metaphone Phonetic Algorithm metric of two strings

TheFuzz.Similarity.DiceSorensen
This module contains functions to calculate the Sorensen-Dice coefficient of 2 given strings

TheFuzz.Similarity.Hamming
This module contains functions to calculate the Hamming distance between 2 given strings

TheFuzz.Similarity.Jaccard
This module contains functions to calculate the Jaccard similarity coefficient between two given strings

TheFuzz.Similarity.Jaro
Calculates the Jaro Distance between two strings

TheFuzz.Similarity.JaroWinkler
Calculates the Jaro-Winkler Distance between two strings

TheFuzz.Similarity.Levenshtein
This module contains functions to calculate the levenshtein distance between given strings

TheFuzz.Similarity.NGram
This module contains functions to calculate the ngram distance between two given strings based on this paper

TheFuzz.Similarity.Overlap
Implements the Overlap Similarity Metric

TheFuzz.Similarity.WeightedLevenshtein
This module contains function to calculate the weighted levenshtein distance between 2 given strings

TheFuzz.StringMetric
Specifies the string metric api which an module needs to implement to provide string comparison methods

TheFuzz.Util
Utilities for TheFuzz